gdi32/tests: Add tests for the advanced graphics mode
The tests are designed to support an automated mode and an interactive
mode.

In automated mode, only tests where the angle used is a multiple of pi/4
are performed to increase the chance of matching images created on Windows.

In interactive test mode, an image is generated for each test for visual
verification.

Signed-off-by: Ralf Habacker <ralf.habacker@freenet.de>